frontend/requirements.txt
# Our how-to docs will use --debug parameter which is only available in Flask 2.2+
Flask>=2.2
flask_session
identity[flask]
msal
requests>=2,<3
# ms_identity_python[flask] @ https://github.com/azure-samples/ms-identity-python/archive/refs/heads/0.9.zip
python-dotenv<0.22
